[Telemetry]
- fixed telemetry
